Players auto-child to boarded ship
Players now become the child of a ship they fly.  Also adjusted the size of a default player's collider because it was too small
Did not fix the movement of a ship not moving players not piloting it as well.  Possibly some transform techniques are needed to work this

In future the ship speed and player speed must be balanced.